Move run model to __main__.py
