+++ wsimod/__main__.py
@@ -1,7 +1,11 @@
 """The entry point for the myproject program."""
 from argparse import ArgumentParser
 from pathlib import Path
+from typing import Any, cast
 
+import pandas as pd
+
+from wsimod.orchestration.model import Model
 from wsimod.validation import validate_io_args
 
 
@@ -29,6 +33,25 @@ def create_parser() -> ArgumentParser:
     )
 
     return parser
+
+
+def run_model(settings: dict[str, Any]) -> None:
+    """Runs the mode with the chosen settings and saves the outputs as csv.
+
+    Args:
+        settings (dict[str, Any]): Settings dictionary with loaded data.
+    """
+    model = Model()
+
+    model.dates = cast(pd.Series, settings["dates"]).drop_duplicates()
+    model.add_nodes(settings["nodes"])
+    model.add_arcs(settings["arcs"])
+
+    flows, tanks, _, surfaces = model.run()
+
+    pd.DataFrame(flows).to_csv(settings["outputs"] / "flows.csv")
+    pd.DataFrame(tanks).to_csv(settings["outputs"] / "tanks.csv")
+    pd.DataFrame(surfaces).to_csv(settings["outputs"] / "surfaces.csv")
 
 
 def run() -> None:
